requirements.txt
pandas
sqlalchemy
python-dotenv
asyncpg
seaborn
numpy